moved the shared schedule utility to vllm_omni/diffusion/sched
