+++ vllm_omni/diffusion/models/dmd2/__init__.py
@@ -3,10 +3,8 @@
 
 from vllm_omni.diffusion.models.dmd2.config import DMD2Config
 from vllm_omni.diffusion.models.dmd2.mixin import DMD2PipelineMixin
-from vllm_omni.diffusion.models.dmd2.schedule import DMD2SigmaSchedule
 
 __all__ = [
     "DMD2Config",
     "DMD2PipelineMixin",
-    "DMD2SigmaSchedule",
 ]
